Remove clients that don't need to be clients
