+++ tck/src/main/java/ee/jakarta/tck/concurrent/api/ManagedExecutors/ManagedExecutorsTests.java
@@ -16,183 +16,324 @@
 
 package ee.jakarta.tck.concurrent.api.ManagedExecutors;
 
-import java.net.URL;
+import java.util.HashMap;
+import java.util.Map;
+import java.util.concurrent.Callable;
+import java.util.concurrent.Executors;
+import java.util.concurrent.Future;
 
 import org.jboss.arquillian.container.test.api.Deployment;
-import org.jboss.arquillian.container.test.api.RunAsClient;
-import org.jboss.arquillian.test.api.ArquillianResource;
 import org.jboss.shrinkwrap.api.ShrinkWrap;
 import org.jboss.shrinkwrap.api.spec.WebArchive;
+import org.junit.jupiter.api.AfterEach;
 import org.junit.jupiter.api.Test;
 
-import ee.jakarta.tck.concurrent.framework.TestClient;
+import ee.jakarta.tck.concurrent.common.managed.task.listener.ListenerEvent;
+import ee.jakarta.tck.concurrent.common.managed.task.listener.ManagedTaskListenerImpl;
+import ee.jakarta.tck.concurrent.common.tasks.CallableTask;
+import ee.jakarta.tck.concurrent.common.tasks.RunnableTask;
+import ee.jakarta.tck.concurrent.framework.TestLogger;
+import ee.jakarta.tck.concurrent.framework.TestUtil;
 import ee.jakarta.tck.concurrent.framework.junit.anno.Common;
 import ee.jakarta.tck.concurrent.framework.junit.anno.Common.PACKAGE;
-import ee.jakarta.tck.concurrent.framework.junit.anno.TestName;
 import ee.jakarta.tck.concurrent.framework.junit.anno.Web;
+import jakarta.enterprise.concurrent.ManagedExecutors;
+import jakarta.enterprise.concurrent.ManagedTask;
 
-@Web @RunAsClient
-@Common({PACKAGE.MANAGED_TASK_LISTENER, PACKAGE.TASKS})
-public class ManagedExecutorsTests extends TestClient {
+@Web
+@Common({ PACKAGE.MANAGED_TASK_LISTENER, PACKAGE.TASKS })
+public class ManagedExecutorsTests {
 
-	@ArquillianResource
-	URL baseURL;
-	
-	//TODO deploy as EJB and JSP artifacts
-	@Deployment(name="ManagedExecutorsTests")
-	public static WebArchive createDeployment() {
-		return ShrinkWrap.create(WebArchive.class)
-				.addPackages(true, ManagedExecutorsTests.class.getPackage())
-				.addAsWebInfResource(ManagedExecutorsTests.class.getPackage(), "web.xml", "web.xml");
-	}
-	
-	@TestName
-	String testname;
-	
-	@Override
-	protected String getServletPath() {
-		return "ManagedExecutorsServlet";
-	}
+    private static final TestLogger log = TestLogger.get(ManagedExecutorsTests.class);
 
-	/*
-	 * @testName: IsCurrentThreadShutdown
-	 * 
-	 * @assertion_ids: CONCURRENCY:JAVADOC:21
-	 * 
-	 * @test_Strategy: Use a regular thread(non-Manageable thread) and verify
-	 * isCurrentThreadShutdown() returns false
-	 * 
-	 */
-	@Test
-	public void IsCurrentThreadShutdown() {
-		runTest(baseURL, testname);
-	}
+    // TODO deploy as EJB and JSP artifacts
+    @Deployment(name = "ManagedExecutorsTests")
+    public static WebArchive createDeployment() {
+        return ShrinkWrap.create(WebArchive.class)
+                .addAsWebInfResource(ManagedExecutorsTests.class.getPackage(), "web.xml", "web.xml");
+    }
 
-	/*
-	 * @testName: IsCurrentThreadShutdown_ManageableThread
-	 * 
-	 * @assertion_ids: CONCURRENCY:JAVADOC:21
-	 * 
-	 * @test_Strategy: Create a ManageableThread from ManagedThreadFactory and check
-	 * the shutdown status.
-	 */
-	@Test
-	public void IsCurrentThreadShutdown_ManageableThread() {
-		runTest(baseURL, testname);
-	}
+    private static final String ENV_ENTRY_JNDI_NAME = "java:comp/env/StringValue";
 
-	/*
-	 * @testName: ManageRunnableTaskWithTaskListener
-	 * 
-	 * @assertion_ids: CONCURRENCY:JAVADOC:22;CONCURRENCY:SPEC:7;
-	 * CONCURRENCY:SPEC:7.1;CONCURRENCY:SPEC:7.2;
-	 * CONCURRENCY:SPEC:4;CONCURRENCY:SPEC:4.2; CONCURRENCY:SPEC:18;
-	 *
-	 * @test_Strategy: Returns a Runnable object that also implements ManagedTask
-	 * interface so it can receive notification of life cycle events with the
-	 * provided ManagedTaskListener when the task is submitted to a
-	 * ManagedExecutorService or a ManagedScheduledExecutorService.
-	 */
-	@Test
-	public void ManageRunnableTaskWithTaskListener() {
-		runTest(baseURL, testname);
-	}
+    private static final String ENV_ENTRY_VALUE = "something";
 
-	/*
-	 * @testName: ManageRunnableTaskWithNullArg
-	 * 
-	 * @assertion_ids: CONCURRENCY:JAVADOC:23
-	 * 
-	 * @test_Strategy: Catch IllegalArgumentException when get the manage task with
-	 * null runnable task.
-	 */
-	@Test
-	public void ManageRunnableTaskWithNullArg() {
-		runTest(baseURL, testname);
-	}
+    private ManagedTaskListenerImpl managedTaskListener = new ManagedTaskListenerImpl();
 
-	/*
-	 * @testName: ManageRunnableTaskWithTaskListenerAndMap
-	 * 
-	 * @assertion_ids: CONCURRENCY:JAVADOC:24;CONCURRENCY:SPEC:13;
-	 * 
-	 * @test_Strategy: Returns a Runnable object that also implements ManagedTask
-	 * interface so it can receive notification of life cycle events with the
-	 * provided ManagedTaskListener and to provide additional execution properties
-	 * when the task is submitted to a ManagedExecutorService
-	 */
-	@Test
-	public void ManageRunnableTaskWithTaskListenerAndMap() {
-		runTest(baseURL, testname);
-	}
+    private boolean shutdown = true;
 
-	/*
-	 * @testName: ManageRunnableTaskWithMapAndNullArg
-	 * 
-	 * @assertion_ids: CONCURRENCY:JAVADOC:25
-	 * 
-	 * @test_Strategy: Catch IllegalArgumentException when get the manage task with
-	 * null runnable task and additional execution properties.
-	 */
-	@Test
-	public void ManageRunnableTaskWithMapAndNullArg() {
-		runTest(baseURL, testname);
-	}
+    @AfterEach
+    public void after() {
+        managedTaskListener.clearEvents();
+    }
 
-	/*
-	 * @testName: ManageCallableTaskWithTaskListener
-	 * 
-	 * @assertion_ids: CONCURRENCY:JAVADOC:26
-	 * 
-	 * @test_Strategy: Returns a Callable object that also implements ManagedTask
-	 * interface so it can receive notification of life cycle events with the
-	 * provided ManagedTaskListener when the task is submitted to a
-	 * ManagedExecutorService
-	 */
-	@Test
-	public void ManageCallableTaskWithTaskListener() {
-		runTest(baseURL, testname);
-	}
+    private RunnableTask createRunnableTask() {
+        return new RunnableTask(ENV_ENTRY_JNDI_NAME, ENV_ENTRY_VALUE, this.getClass().getName());
+    }
 
-	/*
-	 * @testName: ManageCallableTaskWithNullArg
-	 * 
-	 * @assertion_ids: CONCURRENCY:JAVADOC:27
-	 * 
-	 * @test_Strategy: Catch IllegalArgumentException when get the manage task with
-	 * null Callable task.
-	 */
-	@Test
-	public void ManageCallableTaskWithNullArg() {
-		runTest(baseURL, testname);
-	}
+    private CallableTask<String> createCallableTask(String expectedReturnValue) {
+        return new CallableTask<String>(ENV_ENTRY_JNDI_NAME, ENV_ENTRY_VALUE, this.getClass().getName(),
+                expectedReturnValue);
+    }
 
-	/*
-	 * @testName: ManageCallableTaskWithTaskListenerAndMap
-	 * 
-	 * @assertion_ids: CONCURRENCY:JAVADOC:28;CONCURRENCY:SPEC:13.1;
-	 * CONCURRENCY:SPEC:45;CONCURRENCY:SPEC:45.1;
-	 * 
-	 * @test_Strategy: Returns a Callable object that also implements ManagedTask
-	 * interface so it can receive notification of life cycle events with the
-	 * provided ManagedTaskListener and to provide additional execution properties
-	 * when the task is submitted to a ManagedExecutorService
-	 */
-	@Test
-	public void ManageCallableTaskWithTaskListenerAndMap() {
-		runTest(baseURL, testname);
-	}
+    private void assertTaskAndListenerComplete(Future<?> future, RunnableTask runnableTask) {
+        TestUtil.waitForTaskComplete(future);
+        assertListenerComplete(runnableTask);
+    }
 
-	/*
-	 * @testName: ManageCallableTaskWithMapAndNullArg
-	 * 
-	 * @assertion_ids: CONCURRENCY:JAVADOC:29
-	 * 
-	 * @test_Strategy: Catch IllegalArgumentException when get the manage task with
-	 * null Callable task and additional execution properties.
-	 */
-	@Test
-	public void ManageCallableTaskWithMapAndNullArg() {
-		runTest(baseURL, testname);
-	}
+    private void assertTaskAndListenerComplete(String expectedResult, Future<String> future,
+            CallableTask<?> callableTask) {
+        String result = TestUtil.waitForTaskComplete(future);
+        if (!expectedResult.endsWith(result))
+            throw new RuntimeException("Task return different value with expected one.");
+        assertListenerComplete(callableTask);
+    }
+
+    private void assertListenerComplete(RunnableTask task) {
+        // wait for the listener run done.
+        TestUtil.waitForListenerComplete(managedTaskListener);
+        // check listener status.
+        if (!(managedTaskListener.eventCalled(ListenerEvent.SUBMITTED)
+                && managedTaskListener.eventCalled(ListenerEvent.STARTING)
+                && managedTaskListener.eventCalled(ListenerEvent.DONE))) {
+            throw new RuntimeException("TaskListener is not completely executed.");
+        }
+    }
+
+    /*
+     * @testName: IsCurrentThreadShutdown
+     * 
+     * @assertion_ids: CONCURRENCY:JAVADOC:21
+     * 
+     * @test_Strategy: Use a regular thread(non-Manageable thread) and verify
+     * isCurrentThreadShutdown() returns false
+     * 
+     */
+    @Test
+    public void IsCurrentThreadShutdown() {
+        Thread createdThread = TestUtil.getManagedThreadFactory().newThread(new Runnable() {
+            @Override
+            public void run() {
+                shutdown = ManagedExecutors.isCurrentThreadShutdown();
+            }
+        });
+        // Executors.newSingleThreadExecutor() uses Executors.defaultThreadFactory()
+        // to create new thread. So the thread used in this test is a non Manageable
+        // Thread.
+        Future<?> future = Executors.newSingleThreadExecutor().submit(createdThread);
+        TestUtil.waitForTaskComplete(future);
+        if (shutdown) {
+            throw new RuntimeException("Failed because shutdown is set to be true when running job");
+        }
+    }
+
+    /*
+     * @testName: IsCurrentThreadShutdown_ManageableThread
+     * 
+     * @assertion_ids: CONCURRENCY:JAVADOC:21
+     * 
+     * @test_Strategy: Create a ManageableThread from ManagedThreadFactory and check
+     * the shutdown status.
+     */
+    @Test
+    public void IsCurrentThreadShutdown_ManageableThread() {
+        Thread createdThread = TestUtil.getManagedThreadFactory().newThread(new Runnable() {
+            @Override
+            public void run() {
+                shutdown = ManagedExecutors.isCurrentThreadShutdown();
+            }
+        });
+        // Executors.newSingleThreadExecutor(managedThreadFactory) uses
+        // ManagedThreadFactory
+        // to create new (Manageable) thread.
+        Future<?> future = Executors.newSingleThreadExecutor(TestUtil.getManagedThreadFactory()).submit(createdThread);
+        TestUtil.waitForTaskComplete(future);
+        if (shutdown) {
+            throw new RuntimeException("Failed because shutdown is set to be true when running job");
+        }
+    }
+
+    /*
+     * @testName: ManageRunnableTaskWithTaskListener
+     * 
+     * @assertion_ids: CONCURRENCY:JAVADOC:22;CONCURRENCY:SPEC:7;
+     * CONCURRENCY:SPEC:7.1;CONCURRENCY:SPEC:7.2;
+     * CONCURRENCY:SPEC:4;CONCURRENCY:SPEC:4.2; CONCURRENCY:SPEC:18;
+     *
+     * @test_Strategy: Returns a Runnable object that also implements ManagedTask
+     * interface so it can receive notification of life cycle events with the
+     * provided ManagedTaskListener when the task is submitted to a
+     * ManagedExecutorService or a ManagedScheduledExecutorService.
+     */
+    @Test
+    public void ManageRunnableTaskWithTaskListener() {
+        RunnableTask runnableTask = createRunnableTask();
+        Runnable taskWithListener = ManagedExecutors.managedTask(runnableTask, managedTaskListener);
+        Future<?> futureResult = TestUtil.getManagedExecutorService().submit(taskWithListener);
+        assertTaskAndListenerComplete(futureResult, runnableTask);
+    }
+
+    /*
+     * @testName: ManageRunnableTaskWithNullArg
+     * 
+     * @assertion_ids: CONCURRENCY:JAVADOC:23
+     * 
+     * @test_Strategy: Catch IllegalArgumentException when get the manage task with
+     * null runnable task.
+     */
+    @Test
+    public void ManageRunnableTaskWithNullArg() {
+        Runnable nullTask = null;
+        try {
+            ManagedExecutors.managedTask(nullTask, managedTaskListener);
+        } catch (IllegalArgumentException e) {
+            return; // expected
+        } catch (Exception e) {
+            log.warning("Unexpected Exception Caught", e);
+        }
+
+        throw new RuntimeException("Failed to get expected exception");
+    }
+
+    /*
+     * @testName: ManageRunnableTaskWithTaskListenerAndMap
+     * 
+     * @assertion_ids: CONCURRENCY:JAVADOC:24;CONCURRENCY:SPEC:13;
+     * 
+     * @test_Strategy: Returns a Runnable object that also implements ManagedTask
+     * interface so it can receive notification of life cycle events with the
+     * provided ManagedTaskListener and to provide additional execution properties
+     * when the task is submitted to a ManagedExecutorService
+     */
+    @Test
+    public void ManageRunnableTaskWithTaskListenerAndMap() {
+        Map<String, String> properties = new HashMap<String, String>();
+        properties.put("key", "value");
+        RunnableTask runnableTask = createRunnableTask();
+        Runnable task = ManagedExecutors.managedTask(runnableTask, properties, managedTaskListener);
+        boolean pass = false;
+        if (task instanceof ManagedTask) {
+            ManagedTask managedTask = (ManagedTask) task;
+            if (managedTask.getExecutionProperties().get("key") != "value")
+                throw new RuntimeException("Failed to get expected property");
+        }
+
+        assertTaskAndListenerComplete(TestUtil.getManagedExecutorService().submit(task), runnableTask);
+    }
+
+    /*
+     * @testName: ManageRunnableTaskWithMapAndNullArg
+     * 
+     * @assertion_ids: CONCURRENCY:JAVADOC:25
+     * 
+     * @test_Strategy: Catch IllegalArgumentException when get the manage task with
+     * null runnable task and additional execution properties.
+     */
+    @Test
+    public void ManageRunnableTaskWithMapAndNullArg() {
+        Runnable nullTask = null;
+        Map<String, String> properties = new HashMap<String, String>();
+        try {
+            ManagedExecutors.managedTask(nullTask, properties, managedTaskListener);
+        } catch (IllegalArgumentException e) {
+            return; // expected
+        } catch (Exception e) {
+            log.warning("Unexpected Exception Caught", e);
+        }
+        throw new RuntimeException("Failed to get expected exception");
+    }
+
+    /*
+     * @testName: ManageCallableTaskWithTaskListener
+     * 
+     * @assertion_ids: CONCURRENCY:JAVADOC:26
+     * 
+     * @test_Strategy: Returns a Callable object that also implements ManagedTask
+     * interface so it can receive notification of life cycle events with the
+     * provided ManagedTaskListener when the task is submitted to a
+     * ManagedExecutorService
+     */
+    @Test
+    public void ManageCallableTaskWithTaskListener() {
+        String expectedResultStr = "expected something";
+        CallableTask<String> callableTask = createCallableTask(expectedResultStr);
+        Callable<String> taskWithListener = ManagedExecutors.managedTask((Callable<String>) callableTask,
+                managedTaskListener);
+        Future<String> futureResult = TestUtil.getManagedExecutorService().submit(taskWithListener);
+        assertTaskAndListenerComplete(expectedResultStr, futureResult, callableTask);
+    }
+
+    /*
+     * @testName: ManageCallableTaskWithNullArg
+     * 
+     * @assertion_ids: CONCURRENCY:JAVADOC:27
+     * 
+     * @test_Strategy: Catch IllegalArgumentException when get the manage task with
+     * null Callable task.
+     */
+    @Test
+    public void ManageCallableTaskWithNullArg() {
+        Callable<?> nullTask = null;
+        try {
+            ManagedExecutors.managedTask(nullTask, managedTaskListener);
+        } catch (IllegalArgumentException e) {
+            return; // expected
+        } catch (Exception e) {
+            log.warning("Unexpected Exception Caught", e);
+        }
+        throw new RuntimeException("Failed to get expected exception");
+    }
+
+    /*
+     * @testName: ManageCallableTaskWithTaskListenerAndMap
+     * 
+     * @assertion_ids: CONCURRENCY:JAVADOC:28;CONCURRENCY:SPEC:13.1;
+     * CONCURRENCY:SPEC:45;CONCURRENCY:SPEC:45.1;
+     * 
+     * @test_Strategy: Returns a Callable object that also implements ManagedTask
+     * interface so it can receive notification of life cycle events with the
+     * provided ManagedTaskListener and to provide additional execution properties
+     * when the task is submitted to a ManagedExecutorService
+     */
+    @Test
+    public void ManageCallableTaskWithTaskListenerAndMap() {
+        Map<String, String> properties = new HashMap<String, String>();
+        properties.put("key", "value");
+        properties.put(ManagedTask.IDENTITY_NAME, "id");
+        String expectedResultStr = "expected something";
+
+        CallableTask<String> callableTask = createCallableTask(expectedResultStr);
+        Callable<String> task = ManagedExecutors.managedTask((Callable<String>) callableTask, properties,
+                managedTaskListener);
+
+        boolean pass = false;
+        if (task instanceof ManagedTask) {
+            ManagedTask managedTask = (ManagedTask) task;
+            if (managedTask.getExecutionProperties().get("key") != "value")
+                throw new RuntimeException("Failed to get expected property");
+        }
+        assertTaskAndListenerComplete(expectedResultStr, TestUtil.getManagedExecutorService().submit(task),
+                callableTask);
+    }
+
+    /*
+     * @testName: ManageCallableTaskWithMapAndNullArg
+     * 
+     * @assertion_ids: CONCURRENCY:JAVADOC:29
+     * 
+     * @test_Strategy: Catch IllegalArgumentException when get the manage task with
+     * null Callable task and additional execution properties.
+     */
+    @Test
+    public void ManageCallableTaskWithMapAndNullArg() {
+        Callable<?> nullTask = null;
+        Map<String, String> properties = new HashMap<String, String>();
+        try {
+            ManagedExecutors.managedTask(nullTask, properties, managedTaskListener);
+        } catch (IllegalArgumentException e) {
+            return; // expected
+        } catch (Exception e) {
+            log.warning("Unexpected Exception Caught", e);
+        }
+        throw new RuntimeException("Failed to get expected exception");
+    }
 }
