+++ tck/src/main/java/ee/jakarta/tck/concurrent/spec/ManagedExecutorService/managed_servlet/forbiddenapi/ForbiddenAPIServletTests.java
@@ -16,101 +16,122 @@
 
 package ee.jakarta.tck.concurrent.spec.ManagedExecutorService.managed_servlet.forbiddenapi;
 
-import java.net.URL;
+import static org.junit.jupiter.api.Assertions.fail;
+
+import java.util.concurrent.TimeUnit;
 
 import org.jboss.arquillian.container.test.api.Deployment;
-import org.jboss.arquillian.container.test.api.RunAsClient;
-import org.jboss.arquillian.test.api.ArquillianResource;
 import org.jboss.shrinkwrap.api.ShrinkWrap;
 import org.jboss.shrinkwrap.api.spec.WebArchive;
 import org.junit.jupiter.api.Test;
 
-import ee.jakarta.tck.concurrent.framework.TestClient;
-import ee.jakarta.tck.concurrent.framework.junit.anno.TestName;
 import ee.jakarta.tck.concurrent.framework.junit.anno.Web;
+import jakarta.annotation.Resource;
+import jakarta.enterprise.concurrent.ManagedExecutorService;
 
-@Web @RunAsClient
-public class ForbiddenAPIServletTests extends TestClient {
+@Web
+public class ForbiddenAPIServletTests {
 
-	@ArquillianResource
-	URL baseURL;
-	
-	@Deployment(name="ForbiddenAPIServletTests")
-	public static WebArchive createDeployment() {
-		return ShrinkWrap.create(WebArchive.class)
-				.addPackages(true,  ForbiddenAPIServletTests.class.getPackage());
-	}
-	
-	@Override
-	protected String getServletPath() {
-		return "ForbiddenServlet";
-	}
-	
-    @TestName
-    String testname;
+    @Deployment(name = "ForbiddenAPIServletTests")
+    public static WebArchive createDeployment() {
+        return ShrinkWrap.create(WebArchive.class);
+    }
 
-	/*
-	 * @testName: testAwaitTermination
-	 * 
-	 * @assertion_ids:
-	 * CONCURRENCY:SPEC:23;CONCURRENCY:SPEC:24;CONCURRENCY:SPEC:24.1;
-	 * 
-	 * @test_Strategy:
-	 */
-	@Test
-	public void testAwaitTermination() {
-		runTest(baseURL, testname);
-	}
+    @Resource
+    public ManagedExecutorService mes;
 
-	/*
-	 * @testName: testIsShutdown
-	 * 
-	 * @assertion_ids:
-	 * CONCURRENCY:SPEC:23;CONCURRENCY:SPEC:24;CONCURRENCY:SPEC:24.2;
-	 * 
-	 * @test_Strategy:
-	 */
-	@Test
-	public void testIsShutdown() {
-		runTest(baseURL, testname);
-	}
+    /*
+     * @testName: testAwaitTermination
+     * 
+     * @assertion_ids:
+     * CONCURRENCY:SPEC:23;CONCURRENCY:SPEC:24;CONCURRENCY:SPEC:24.1;
+     * 
+     * @test_Strategy:
+     */
+    @Test
+    public void testAwaitTermination() {
+        try {
+            mes.awaitTermination(10, TimeUnit.SECONDS);
+        } catch (IllegalStateException e) { // what expected.
+            return;
+        } catch (Exception e) {
+            fail(e.toString());
+        }
+    }
 
-	/*
-	 * @testName: testIsTerminated
-	 * 
-	 * @assertion_ids:
-	 * CONCURRENCY:SPEC:23;CONCURRENCY:SPEC:24;CONCURRENCY:SPEC:24.3;
-	 * 
-	 * @test_Strategy:
-	 */
-	@Test
-	public void testIsTerminated() {
-		runTest(baseURL, testname);
-	}
+    /*
+     * @testName: testIsShutdown
+     * 
+     * @assertion_ids:
+     * CONCURRENCY:SPEC:23;CONCURRENCY:SPEC:24;CONCURRENCY:SPEC:24.2;
+     * 
+     * @test_Strategy:
+     */
+    @Test
+    public void testIsShutdown() {
+        try {
+            mes.isShutdown();
+        } catch (IllegalStateException e) { // what expected
+            return;
+        } catch (Exception e) {
+            fail(e.toString());
+        }
+    }
 
-	/*
-	 * @testName: testShutdown
-	 * 
-	 * @assertion_ids:
-	 * CONCURRENCY:SPEC:23;CONCURRENCY:SPEC:24;CONCURRENCY:SPEC:24.4;
-	 * 
-	 * @test_Strategy:
-	 */
-	@Test
-	public void testShutdown() {
-		runTest(baseURL, testname);
-	}
+    /*
+     * @testName: testIsTerminated
+     * 
+     * @assertion_ids:
+     * CONCURRENCY:SPEC:23;CONCURRENCY:SPEC:24;CONCURRENCY:SPEC:24.3;
+     * 
+     * @test_Strategy:
+     */
+    @Test
+    public void testIsTerminated() {
+        try {
+            mes.isTerminated();
+        } catch (IllegalStateException e) { // what expected
+            return;
+        } catch (Exception e) {
+            fail(e.toString());
+        }
+    }
 
-	/*
-	 * @testName: testShutdownNow
-	 * 
-	 * @assertion_ids:
-	 * CONCURRENCY:SPEC:23;CONCURRENCY:SPEC:24;CONCURRENCY:SPEC:24.5;
-	 * 
-	 * @test_Strategy:
-	 */
-	@Test
-	public void testShutdownNow() {
-		runTest(baseURL, testname);
-	}
+    /*
+     * @testName: testShutdown
+     * 
+     * @assertion_ids:
+     * CONCURRENCY:SPEC:23;CONCURRENCY:SPEC:24;CONCURRENCY:SPEC:24.4;
+     * 
+     * @test_Strategy:
+     */
+    @Test
+    public void testShutdown() {
+        try {
+            mes.shutdown();
+        } catch (IllegalStateException e) { // what expected
+            return;
+        } catch (Exception e) {
+            fail(e.toString());
+        }
+    }
+
+    /*
+     * @testName: testShutdownNow
+     * 
+     * @assertion_ids:
+     * CONCURRENCY:SPEC:23;CONCURRENCY:SPEC:24;CONCURRENCY:SPEC:24.5;
+     * 
+     * @test_Strategy:
+     */
+    @Test
+    public void testShutdownNow() {
+        try {
+            mes.shutdownNow();
+        } catch (IllegalStateException e) { // what expected
+            return;
+        } catch (Exception e) {
+            fail(e.toString());
+        }
+    }
 }
