+++ tck/src/main/java/ee/jakarta/tck/concurrent/spec/ManagedScheduledExecutorService/inheritedapi_servlet/InheritedAPIServletTests.java
@@ -16,124 +16,254 @@
 
 package ee.jakarta.tck.concurrent.spec.ManagedScheduledExecutorService.inheritedapi_servlet;
 
-import java.net.URL;
+import static org.junit.jupiter.api.Assertions.assertEquals;
+import static org.junit.jupiter.api.Assertions.fail;
+
+import java.util.ArrayList;
+import java.util.List;
+import java.util.concurrent.CancellationException;
+import java.util.concurrent.Future;
+import java.util.concurrent.ScheduledFuture;
+import java.util.concurrent.TimeUnit;
+import java.util.concurrent.TimeoutException;
 
 import org.jboss.arquillian.container.test.api.Deployment;
-import org.jboss.arquillian.container.test.api.RunAsClient;
-import org.jboss.arquillian.test.api.ArquillianResource;
 import org.jboss.shrinkwrap.api.ShrinkWrap;
 import org.jboss.shrinkwrap.api.spec.WebArchive;
+import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 
-import ee.jakarta.tck.concurrent.framework.TestClient;
+import ee.jakarta.tck.concurrent.common.fixed.counter.CounterRunnableTask;
+import ee.jakarta.tck.concurrent.common.fixed.counter.StaticCounter;
+import ee.jakarta.tck.concurrent.common.tasks.CommonTasks;
+import ee.jakarta.tck.concurrent.framework.TestConstants;
+import ee.jakarta.tck.concurrent.framework.TestUtil;
 import ee.jakarta.tck.concurrent.framework.junit.anno.Common;
 import ee.jakarta.tck.concurrent.framework.junit.anno.Common.PACKAGE;
-import ee.jakarta.tck.concurrent.framework.junit.anno.TestName;
 import ee.jakarta.tck.concurrent.framework.junit.anno.Web;
 
-@Web @RunAsClient
-@Common({PACKAGE.TASKS, PACKAGE.FIXED_COUNTER})
-public class InheritedAPIServletTests extends TestClient {
-	
-	@ArquillianResource
-	URL baseURL;
-	
-	@Deployment(name="InheritedAPIServletTests")
-	public static WebArchive createDeployment() {
-		return ShrinkWrap.create(WebArchive.class)
-				.addPackages(true, InheritedAPIServletTests.class.getPackage());
-	}
-	
-    @TestName
-    String testname;
-	
-	@Override
-	protected String getServletPath() {
-		return "InheritedAPIServlet";
-	}
+@Web
+@Common({ PACKAGE.TASKS, PACKAGE.FIXED_COUNTER })
+public class InheritedAPIServletTests {
 
-	/*
-	 * @testName: testApiSubmit
-	 * 
-	 * @assertion_ids: CONCURRENCY:SPEC:44.1
-	 * 
-	 * @test_Strategy:
-	 */
-	@Test
-	public void testApiSubmit() {
-		runTest(baseURL, testname);
-	}
+    @Deployment(name = "InheritedAPIServletTests")
+    public static WebArchive createDeployment() {
+        return ShrinkWrap.create(WebArchive.class);
+    }
 
-	/*
-	 * @testName: testApiExecute
-	 * 
-	 * @assertion_ids: CONCURRENCY:SPEC:44.2
-	 * 
-	 * @test_Strategy:
-	 */
-	@Test
-	public void testApiExecute() {
-		runTest(baseURL, testname);
-	}
+    @BeforeEach
+    public void before() {
+        StaticCounter.reset();
+    }
 
-	/*
-	 * @testName: testApiInvokeAll
-	 * 
-	 * @assertion_ids: CONCURRENCY:SPEC:44.3
-	 * 
-	 * @test_Strategy:
-	 */
-	@Test
-	public void testApiInvokeAll() {
-		runTest(baseURL, testname);
-	}
+    /*
+     * @testName: testApiSubmit
+     * 
+     * @assertion_ids: CONCURRENCY:SPEC:44.1
+     * 
+     * @test_Strategy:
+     */
+    @Test
+    public void testApiSubmit() throws Exception {
+        Future result = TestUtil.getManagedScheduledExecutorService().submit(new CommonTasks.SimpleCallable());
+        TestUtil.waitTillFutureIsDone(result);
+        assertEquals(result.get(), CommonTasks.SIMPLE_RETURN_STRING);
 
-	/*
-	 * @testName: testApiInvokeAny
-	 * 
-	 * @assertion_ids: CONCURRENCY:SPEC:44.4
-	 * 
-	 * @test_Strategy:
-	 */
-	@Test
-	public void testApiInvokeAny() {
-		runTest(baseURL, testname);
-	}
+        result = TestUtil.getManagedScheduledExecutorService().submit(new CommonTasks.SimpleRunnable());
+        TestUtil.waitTillFutureIsDone(result);
+        result.get();
 
-	/*
-	 * @testName: testApiSchedule
-	 * 
-	 * @assertion_ids: CONCURRENCY:SPEC:44.5
-	 * 
-	 * @test_Strategy:
-	 */
-	@Test
-	public void testApiSchedule() {
-		runTest(baseURL, testname);
-	}
+        result = TestUtil.getManagedScheduledExecutorService().submit(new CommonTasks.SimpleRunnable(),
+                CommonTasks.SIMPLE_RETURN_STRING);
+        TestUtil.waitTillFutureIsDone(result);
+        assertEquals(result.get(), CommonTasks.SIMPLE_RETURN_STRING);
+    }
 
-	/*
-	 * @testName: testApiScheduleAtFixedRate
-	 * 
-	 * @assertion_ids: CONCURRENCY:SPEC:44.6
-	 * 
-	 * @test_Strategy:
-	 */
-	@Test
-	public void testApiScheduleAtFixedRate() {
-		runTest(baseURL, testname);
-	}
+    /*
+     * @testName: testApiExecute
+     * 
+     * @assertion_ids: CONCURRENCY:SPEC:44.2
+     * 
+     * @test_Strategy:
+     */
+    @Test
+    public void testApiExecute() {
+        TestUtil.getManagedScheduledExecutorService().execute(new CounterRunnableTask());
+        waitForCounter(1);
+    }
 
-	/*
-	 * @testName: testApiScheduleWithFixedDelay
-	 * 
-	 * @assertion_ids: CONCURRENCY:SPEC:44.7
-	 * 
-	 * @test_Strategy:
-	 */
-	@Test
-	public void testApiScheduleWithFixedDelay() {
-		runTest(baseURL, testname);
-	}
+    /*
+     * @testName: testApiInvokeAll
+     * 
+     * @assertion_ids: CONCURRENCY:SPEC:44.3
+     * 
+     * @test_Strategy:
+     */
+    @Test
+    public void testApiInvokeAll() throws Exception {
+        List taskList = new ArrayList();
+        taskList.add(new CommonTasks.SimpleArgCallable(1));
+        taskList.add(new CommonTasks.SimpleArgCallable(2));
+        taskList.add(new CommonTasks.SimpleArgCallable(3));
+        List<Future> resultList = TestUtil.getManagedScheduledExecutorService().invokeAll(taskList);
+        for (Future each : resultList) {
+            TestUtil.waitTillFutureIsDone(each);
+        }
+        assertEquals(resultList.get(0).get(), 1);
+        assertEquals(resultList.get(1).get(), 2);
+        assertEquals(resultList.get(2).get(), 3);
+        resultList = TestUtil.getManagedScheduledExecutorService().invokeAll(taskList,
+                TestConstants.WaitTimeout.getSeconds(), TimeUnit.SECONDS);
+        for (Future each : resultList) {
+            TestUtil.waitTillFutureIsDone(each);
+        }
+        assertEquals(resultList.get(0).get(), 1);
+        assertEquals(resultList.get(1).get(), 2);
+        assertEquals(resultList.get(2).get(), 3);
+
+        try {
+            taskList = new ArrayList();
+            taskList.add(new CommonTasks.SimpleCallable(TestConstants.WaitTimeout.toMillis()));
+            taskList.add(new CommonTasks.SimpleCallable(TestConstants.WaitTimeout.toMillis()));
+            resultList = TestUtil.getManagedScheduledExecutorService().invokeAll(taskList,
+                    TestConstants.PollInterval.getSeconds(), TimeUnit.SECONDS);
+            for (Future each : resultList) {
+                TestUtil.waitTillFutureThrowsException(each, CancellationException.class);
+            }
+        } catch (Exception ex) {
+            fail(ex.getMessage());
+        }
+    }
+
+    /*
+     * @testName: testApiInvokeAny
+     * 
+     * @assertion_ids: CONCURRENCY:SPEC:44.4
+     * 
+     * @test_Strategy:
+     */
+    @Test
+    public void testApiInvokeAny() throws Exception {
+        List taskList = new ArrayList();
+        taskList.add(new CommonTasks.SimpleArgCallable(1));
+        taskList.add(new CommonTasks.SimpleArgCallable(2));
+        taskList.add(new CommonTasks.SimpleArgCallable(3));
+        Object result = TestUtil.getManagedScheduledExecutorService().invokeAny(taskList);
+        TestUtil.assertInRange(new Integer[] { 1, 2, 3 }, result);
+        result = TestUtil.getManagedScheduledExecutorService().invokeAny(taskList,
+                TestConstants.WaitTimeout.getSeconds(), TimeUnit.SECONDS);
+        TestUtil.assertInRange(new Integer[] { 1, 2, 3 }, result);
+
+        try {
+            taskList = new ArrayList();
+            taskList.add(new CommonTasks.SimpleCallable(TestConstants.WaitTimeout.toMillis()));
+            taskList.add(new CommonTasks.SimpleCallable(TestConstants.WaitTimeout.toMillis()));
+            result = TestUtil.getManagedScheduledExecutorService().invokeAny(taskList,
+                    TestConstants.PollInterval.getSeconds(), TimeUnit.SECONDS);
+        } catch (TimeoutException e) {
+            return; // expected
+        } catch (Exception ex) {
+            fail(ex.getMessage());
+        }
+        fail("Task should be cancelled because of timeout");
+    }
+
+    /*
+     * @testName: testApiSchedule
+     * 
+     * @assertion_ids: CONCURRENCY:SPEC:44.5
+     * 
+     * @test_Strategy:
+     */
+    @Test
+    public void testApiSchedule() throws Exception {
+        Future result = TestUtil.getManagedScheduledExecutorService().schedule(new CommonTasks.SimpleCallable(),
+                TestConstants.PollInterval.getSeconds(), TimeUnit.SECONDS);
+        TestUtil.waitTillFutureIsDone(result);
+        assertEquals(result.get(), CommonTasks.SIMPLE_RETURN_STRING);
+
+        result = TestUtil.getManagedScheduledExecutorService().schedule(new CommonTasks.SimpleRunnable(),
+                TestConstants.PollInterval.getSeconds(), TimeUnit.SECONDS);
+        TestUtil.waitTillFutureIsDone(result);
+        assertEquals(result.get(), null);
+    }
+
+    /*
+     * @testName: testApiScheduleAtFixedRate
+     * 
+     * @assertion_ids: CONCURRENCY:SPEC:44.6
+     * 
+     * @test_Strategy:
+     */
+    @Test
+    public void testApiScheduleAtFixedRate() {
+        ScheduledFuture result = null;
+
+        try {
+            result = TestUtil.getManagedScheduledExecutorService().scheduleAtFixedRate(new CounterRunnableTask(),
+                    TestConstants.PollInterval.getSeconds(), TestConstants.PollInterval.getSeconds(), TimeUnit.SECONDS);
+            TestUtil.sleep(TestConstants.WaitTimeout);
+            TestUtil.assertIntInRange(TestConstants.PollsPerTimeout - 2, TestConstants.PollsPerTimeout + 2,
+                    StaticCounter.getCount());
+        } catch (Exception e) {
+            fail(e.getMessage());
+        } finally {
+            if (result != null) {
+                result.cancel(true);
+                // Sleep to ensure cancel take effect.
+                try {
+                    TestUtil.sleep(TestConstants.PollInterval);
+                } catch (Exception e) {
+                }
+            }
+        }
+    }
+
+    /*
+     * @testName: testApiScheduleWithFixedDelay
+     * 
+     * @assertion_ids: CONCURRENCY:SPEC:44.7
+     * 
+     * @test_Strategy:
+     */
+    @Test
+    public void testApiScheduleWithFixedDelay() {
+        ScheduledFuture result = null;
+        try {
+            result = TestUtil.getManagedScheduledExecutorService().scheduleWithFixedDelay(
+                    new CounterRunnableTask(TestConstants.PollInterval.toMillis()),
+                    TestConstants.PollInterval.getSeconds(), TestConstants.PollInterval.getSeconds(), TimeUnit.SECONDS);
+            TestUtil.sleep(TestConstants.WaitTimeout);
+            TestUtil.assertIntInRange((TestConstants.PollsPerTimeout / 2) - 2, (TestConstants.PollsPerTimeout / 2) + 2,
+                    StaticCounter.getCount());
+        } catch (Exception e) {
+            fail(e.getMessage());
+        } finally {
+            if (result != null) {
+                result.cancel(true);
+                // Sleep to ensure cancel take effect.
+                try {
+                    TestUtil.sleep(TestConstants.PollInterval);
+                } catch (Exception e) {
+                }
+            }
+        }
+    }
+
+    private void waitForCounter(int expected) {
+        long start = System.currentTimeMillis();
+
+        while (expected != StaticCounter.getCount()) {
+            try {
+                TestUtil.sleep(TestConstants.PollInterval);
+            } catch (InterruptedException ignore) {
+            }
+
+            if ((System.currentTimeMillis() - start) > TestConstants.WaitTimeout.toMillis()) {
+                throw new RuntimeException("Static counter did not produce expected counter before timeout. Expected: "
+                        + expected + " Actual: " + StaticCounter.getCount());
+            }
+        }
+    }
 
 }
