+++ tck/src/main/java/ee/jakarta/tck/concurrent/spec/ManagedThreadFactory/tx/TransactionTests.java
@@ -30,7 +30,7 @@ import ee.jakarta.tck.concurrent.framework.TestClient;
 import ee.jakarta.tck.concurrent.framework.URLBuilder;
 import ee.jakarta.tck.concurrent.framework.junit.anno.Web;
 
-@Web @RunAsClient
+@Web @RunAsClient //Requires client testing due to annotation configuration
 public class TransactionTests extends TestClient {
 
 	@ArquillianResource
